+++ manage.py
@@ -1,6 +1,7 @@
 #! ./.venv/bin/python
 
 """Django's command-line utility for administrative tasks."""
+
 import os
 import sys